Create CloneInitializer.cs
Avoid creating clones of other clones. The only spawn point of a clone should be the location of the original object. 

This should use a separate empty GameObject to attach CloneInitializer.cs to. 

Don't attach this file to the objects you are cloning. This will cause clones to create clones indefinitely and exponentially causing unity to crash.